# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for site",
        "title": "Backend API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/cart": {
            "get": {
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "description": "Get user cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Get user cart",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CartItemsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "description": "Post product in cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Post product in cart",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cart.PostProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CartItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete product in user cart with ids",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cart"
                ],
                "summary": "Delete product in cart",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "cart.PostProductRequest": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "model.CartItem": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.CartItemsResponse": {
            "type": "object",
            "properties": {
                "product": {
                    "$ref": "#/definitions/model.Product"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Product": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "business_id": {
                    "type": "integer"
                },
                "category": {
                    "$ref": "#/definitions/model.ProductCategory"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "discount": {
                    "type": "number"
                },
                "estimated_delivery": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "description": "Загружается отдельно",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProductImage"
                    }
                },
                "price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "related_products": {
                    "description": "Загружается отдельно",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "review_count": {
                    "type": "integer"
                },
                "reviews": {
                    "description": "Загружается отдельно",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProductReview"
                    }
                },
                "sku": {
                    "type": "string"
                },
                "specifications": {
                    "description": "Загружается отдельно",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ProductSpecification"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.ProductCategory": {
            "type": "string",
            "enum": [
                "ELECTRONICS",
                "HOME",
                "FASHION",
                "SPORTS",
                "BEAUTY",
                "TOYS",
                "BOOKS",
                "FOOD",
                "OTHER"
            ],
            "x-enum-varnames": [
                "ProductCategoryElectronics",
                "ProductCategoryHome",
                "ProductCategoryFashion",
                "ProductCategorySports",
                "ProductCategoryBeauty",
                "ProductCategoryToys",
                "ProductCategoryBooks",
                "ProductCategoryFood",
                "ProductCategoryOther"
            ]
        },
...
        "model.ProductImage": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_uuid": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_primary": {
                    "type": "boolean"
                },
                "product_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "description": "Не хранится в базе данных напрямую",
                    "type": "string"
                }
            }
        },
        "model.ProductReview": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "description": "Не хранится в базе данных напрямую",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "product_id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "model.ProductSpecification": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "product_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2PasswordBearer": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/user/token",
            "scopes": {
                "read": "Grants read access",
                "write": "Grants write access"
            }
        }
    }
}